pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_ops"
version = "0.0.1" # major.minor.patch
description = "Repository for the DTU Machine Learning Operations course."
authors = [
  { name = "Vincent Van Schependom", email = "vincent.van.schependom@gmail.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc[ssh]>=3.65.0",
    "fastapi==0.115.6",
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "matplotlib>=3.10.8",
    "pathspec<1.0.0",
    "scikit-learn>=1.8.0",
    "torch==2.9.1",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "torchvision>=0.24.1",
    "loguru>=0.7.3",
    "wandb>=0.23.1",
]

# Optional dependency groups, which can be installed using:
#    uv sync --group dev
# To install all groups, use:
#    uv sync --all-groups
[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "mkdocstrings-python==1.12.2",
    "mypy>=1.19.1", # static type checker
    "ruff>=0.14.10",
    "ty>=0.0.9",
    "snakeviz>=2.2.2",
    "torch-tb-profiler>=0.4.3",
    "tensorboard-plugin-profile>=2.21.3",
]

[project.scripts]
train = "ml_ops.train:main" # creates an executable `train` script that calls src/ml_ops/train.py:main()

## To declare which optional dependencies should be installed when
## executing `uv sync`, use the following section.
# [tool.uv]
# default-groups = ["dev"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
lint.select = ["I"] # Import-related rules
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]

## If we created our project with
##    uv init --lib,
## then these lines would be added:
# [build-system]
# requires = ["uv_build>=0.9.21,<0.10.0"]
# build-backend = "uv_build"
